% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_custom_words.R
\name{add_custom_words}
\alias{add_custom_words}
\title{Add Custom Words}
\usage{
add_custom_words(words)
}
\arguments{
\item{words}{string or character vector, words to be added}
}
\value{

}
\description{
Add Custom Words
}
\examples{
add_custom_words("\u662f\u5481\u7684")
}
