% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens_cantonese.R
\name{tokens_cantonese}
\alias{tokens_cantonese}
\title{Tokenise Cantonese texts}
\usage{
tokens_cantonese(x)
}
\arguments{
\item{x}{string, character vector, or quanteda corpus object}
}
\value{
a tokens object
}
\description{
This function tokenises Cantonese texts and create a tokens object.
}
\examples{
\dontrun{
corpus \%>\% tokens_cantonese() \%>\% dfm()
}
}
